backend/requirements.txt
# ResolveX Backend - FastAPI + MySQL
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email-validator]==2.5.3
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7
apscheduler==3.10.4
python-dotenv==1.0.0
aiofiles==23.2.1
